Refactored EvacuationReportService and EvacuationHelper for improved phase and building size Logic

+++ src/main/java/nl/novi/eindopdracht_cursusadministratie/model/report/EvacuationReport.java
@@ -33,7 +33,7 @@ public class EvacuationReport {
     @Enumerated(EnumType.STRING)
     private EvacuationPhase phase;
 
-    private double evacuationTimeMinutes;
+    private int evacuationTimeMinutes;
     private String buildingSize;
 
     private String observations;
